Remove AccountLoader and integrate into AccountManager
